Script apaga versão feito

+++ apaga.sh
@@ -1,7 +1,12 @@
 #!/bin/bash
 
-# TESTE
+# sh apaga.sh "0.0.1" # vai apagar a versao 0.0.1
 
-echo "Apagando versão $1..."
-rm -r "versioned_docs/version-$1"
-rm -r "versioned_sidebars/version-$1-sidebars.json"
+if [ -e "versioned_docs/version-$1" ] ; then
+	echo "Apagando versão $1..."
+	rm -r "versioned_docs/version-$1"
+	rm -r "versioned_sidebars/version-$1-sidebars.json"
+	vim versions.json
+else
+	echo "Essa versão não existe, ou já apada."
+fi
